Fix various compilation warnings

+++ Code/Core/FileIO/FileIO.cpp
@@ -175,7 +175,7 @@
             }
 
             // try to remove read-only flag on dst file
-            dwAttrs = ( dwAttrs & ~FILE_ATTRIBUTE_READONLY );
+            dwAttrs = ( dwAttrs & (uint32_t)(~FILE_ATTRIBUTE_READONLY) );
             if ( FALSE == SetFileAttributes( dstFileName, dwAttrs ) )
             {
                 return false; // failed to remove read-only flag
@@ -760,7 +760,7 @@
 
         // determine the new attributes
         DWORD dwNewAttrs = ( readOnly ) ? ( dwAttrs | FILE_ATTRIBUTE_READONLY )
-                                        : ( dwAttrs & ~FILE_ATTRIBUTE_READONLY );
+                                        : ( dwAttrs & (uint32_t)(~FILE_ATTRIBUTE_READONLY) );
 
         // nothing to do if they are the same
         if ( dwNewAttrs == dwAttrs )
